+++ src/Examples/FileExtensionExample.cs
@@ -8,18 +8,17 @@ public static class FileExtensionExample
     {
         const string path = @"C:\Users\Tester\Desktop\Interesting Text.odt";
 
-        string extension = Path.GetExtension(path);
-        string mimeType = MimeType.FromFileExtension(extension);
+        string mimeType = MimeConverter.FromFileName(path);
 
-        Console.Write($"The MIME type for \"{extension}\" is: ");
+        Console.Write($"The MIME type for \"{path}\" is: ");
         Console.WriteLine(mimeType);
         Console.Write("The file type extension for this MIME type is: ");
-        Console.WriteLine(MimeType.ToFileExtension(mimeType));
+        Console.WriteLine(MimeConverter.ToFileTypeExtension(mimeType));
     }
 }
 /*
 Console Output:
 
-The MIME type for ".odt" is: application/vnd.oasis.opendocument.text
+The MIME type for "C:\Users\Tester\Desktop\Interesting Text.odt" is: application/vnd.oasis.opendocument.text
 The file type extension for this MIME type is: .odt
  */