+++ src/FolkerKinzel.MimeTypes/md/ReleaseNotes.md
@@ -5,12 +5,15 @@ You may be wondering why a major version is being released as a preview again. I
 Extensive changes to the public interface were required to fix this issue. I think the best way to deal with this was to release the changes as soon as possible to upset as few users as possible. I apologize.
 
 The most important changes are:
-- The `MimeTypeStruct` has been renamed to `MimeTypeInfo` in order to show its use more clearly.
-- `MimeType` is a static class now that works on strings and has the task to convert file type extensions into Internet Media Types and vice versa.
+- The `MimeType` struct has been renamed to `MimeTypeInfo` in order to show its use more clearly.
+- `MimeType` is a static class that works on strings and has the task to convert file names into Internet Media Types and Internet Media Types into appropriate file type extensions.
+- The method `MimeTypeBuilder.Build()` has been renamed to `MimeTypeBuilder.AsInfo()` because the MimeTypeInfo struct is directly used only in special cases.
 - `MimeTypeBuilder` got additional methods:
 ```csharp
 MimeTypeBuilder Create(in MimeTypeInfo info);
 MimeTypeBuilder RemoveParameter(string key);
+string ToString(MimeFormats, int);
+StringBuilder AppendTo(StringBuilder, MimeFormats, int);
 ```
 
 >**Project reference:** On some systems, the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and check the "Allow" checkbox - if it is present - in the lower right corner of the General tab in the Properties dialog.